requirements.txt
nselib
mysql-connector-python